Begin safe parse lib
